put regexp and literal parse TESTS in one file
